custom prompt
fixed a few errors from the build_propmt method that I pushed last time. I added a few cases after I tested for the the remainder of the custom prompt things, but didn't compile test them before pushing.

also added a 'prompt' command to the built-in section. don't forget to surround the new prompt with ""